refactor: delete useMessagePreviewHeaderNavigation asd move logic
